synbiobot_backup_benchling/requirements.txt
pyairtable==1.4.0
validators==0.18.2
benchlingapi==2.1.12
biopython==1.78
natsort==8.4.0
dateutils==0.6.12
python-dotenv==0.21.0
